requirements.txt
requests
schedule
python-dotenv
openai>=1.1.0
robin_stocks